% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewin.R
\name{firewin}
\alias{firewin}
\title{firewin()}
\usage{
firewin(
  firecaps = "default",
  noimage = TRUE,
  killjava = TRUE,
  killfire = TRUE,
  killports = FALSE,
  port = 4567,
  waitsecmin = 2,
  waitsecmax = 3
)
}
\arguments{
\item{firecaps}{firefox capabilities default=default, other options dtrackprotect dtrackprotectnofinger dtrackprotectdonot dtrackprotectdonotnofinger}

\item{noimage}{block images? default=TRUE}

\item{killjava}{kill java processes prior to launching, default=TRUE}

\item{killfire}{kill firefox processes prior to launching, default=TRUE}

\item{killports}{kill port set to use prior to connecting, default=FALSE}

\item{port}{port to use while connecting}

\item{waitsecmin}{optional min number of seconds to wait after connecting}

\item{waitsecmax}{optional max number of seconds to wait after connecting}
}
\description{
launch windows firefox
}
