% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNORD.R
\name{getNORD}
\alias{getNORD}
\title{getNORD}
\usage{
getNORD(state)
}
\arguments{
\item{state}{Name of state as two char capital eg "FL" or first character capital state name ie "Florida"}
}
\description{
Get Nord server for specified location
}
