% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getua.R
\name{getua}
\alias{getua}
\title{getua()}
\usage{
getua(
  brow = "random",
  browver = "random",
  uasys = "random",
  forcenew = FALSE,
  mainbrow = TRUE
)
}
\arguments{
\item{brow}{browser type, one of firefox, safari, or chrome}

\item{browver}{string of browser version to use}

\item{uasys}{string of system type to use, one of windows, mac/macintosh, linux}

\item{forcenew}{true or false force a new update of user agents to be downloaded and used}

\item{mainbrow}{use only firefox, chrome, or safari uagents}
}
\description{
get user agent based on specs or random
}
