% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renord.R
\name{renord}
\alias{renord}
\title{renord()}
\usage{
renord(nordloc = "randomcity", waitsec = 4)
}
\arguments{
\item{nordloc}{nord location to connect to, default="random"}

\item{waitsec}{optional number of seconds to wait after connecting}
}
\description{
redo nordvpn connection
}
