% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renordwin.R
\name{renordwin}
\alias{renordwin}
\title{renordwin()}
\usage{
renordwin(
  waitsecmin = 8,
  waitsecmax = 10,
  nordloc = "randomUS",
  disconnect = FALSE
)
}
\arguments{
\item{waitsecmin}{optional min number of seconds to wait after connecting}

\item{waitsecmax}{optional max number of seconds to wait after connecting}

\item{nordloc}{nord location to connect to, default="randomUS"}

\item{disconnect}{set TRUE to disconnect from nord before reconnecting}
}
\description{
redo nordvpn connection
}
